requirements.txt
Django==3.2
psycopg2==2.9.2
djangorestframework==3.13.1
django-filter==22.1
pytest-django==4.5.2
pytest==6.2.5
django-simple-history==3.1.1
django-debug-toolbar==3.4.0